fixBook

+++ BE/app/controllers/book.controller.js
@@ -3,7 +3,7 @@ const Book = require("../model/book.model");
 const ApiError = require('../middleware/api-error');
 
 
-exports.getAllBooks = async (req, res) => {
+exports.getAllBooks = async (req, res,next) => {
     try {
         const books = await Book.find({});
         return res.status(200).json(books);
@@ -11,9 +11,9 @@ exports.getAllBooks = async (req, res) => {
         return res.status(500).json({ message: "Internal Server Error" });
     }
 };
-exports.getBook = async (req, res) => {
+exports.getBook = async (req, res,next) => {
     const bookId = req.params.id;
-
+    if (!mongoose.Types.ObjectId.isValid(bookId)) return next(new ApiError(400, 'Book not found'));
     try {
         const book = await Book.findById(bookId);
         if (!book) {
@@ -44,31 +44,38 @@ exports.createBook = async (req, res, next) => {
 };
 
 
-exports.updateBook = async (req, res) => {
-    const bookId = req.params.id;
-    const updateFields = req.body;
-
+exports.updateBook = async (req, res,next) => {
     try {
-        const updatedBook = await Book.findByIdAndUpdate(bookId, updateFields, { new: true });
-        if (!updatedBook) {
-            return res.status(404).json({ message: "Book not found" });
+        const bookId = req.params.id;
+        const { title, price, publisher, publicationDate, available } = req.body;
+        const isExistingBook = await Book.findOne({ title: { $regex: new RegExp(title, 'i') } });
+
+        if (isExistingBook) {
+            return next(new ApiError(400, "Sách này đã tồn tại Hoặc trùng với tên cũ"));
         }
-        return res.status(200).json(updatedBook);
+
+        const updatedBook = await Book.findByIdAndUpdate(bookId, { title, price, publisher, publicationDate, available } );
+
+        if (!updatedBook) {
+            return next(new ApiError(500, "KHÔNG THỂ CẬP NHẬT SÁCH NÀY."));
+        }
+        res.status(201).send({ message: "success" })
     } catch (error) {
-        return res.status(500).json({ message: "Internal Server Error" });
+            // console.error(error);
+        return next(new ApiError(500, "Internal Server Error"));
     }
 };
 
 
-exports.deleteBook = async (req, res) => {
+exports.deleteBook = async (req, res,next) => {
     const bookId = req.params.id;
-
+    if (!mongoose.Types.ObjectId.isValid(bookId)) return next(new ApiError(401, 'Không tìm thấy sách này.'));
     try {
         const deletedBook = await Book.findByIdAndDelete(bookId);
         if (!deletedBook) {
-            return res.status(404).json({ message: "Book not found" });
+            return res.status(404).json({ message: 'Không tìm thấy sách này.' });
         }
-        return res.status(200).json({ message: "Book deleted successfully" });
+        return res.status(201).json({ message: "success" });
     } catch (error) {
         return res.status(500).json({ message: "Internal Server Error" });
     }
